tests: arch: arm: move test arm_runtime_nmi to new ztest API Migrate the testsuite tests/arch/arm/arm_runtime_nmi to the new ztest API.
