+++ tests/arch/arm/arm_runtime_nmi/prj.conf
@@ -1,2 +1,3 @@
 CONFIG_RUNTIME_NMI=y
 CONFIG_ZTEST=y
+CONFIG_ZTEST_NEW_API=y
